RavenDB-19091 refactoring
